upload images, display images on Water Gardens page and on individual garden pages.
