+++ src/main/java/org/launchcode/water_garden_tour/controllers/GardenController.java
@@ -1,37 +1,31 @@
 package org.launchcode.water_garden_tour.controllers;
 
+import org.apache.tomcat.util.codec.binary.Base64;
+import org.apache.tomcat.util.http.fileupload.IOUtils;
+import org.jvnet.staxex.Base64EncoderStream;
 import org.launchcode.water_garden_tour.models.data.FeatureRepository;
 import org.launchcode.water_garden_tour.models.data.GardenRepository;
-import org.launchcode.water_garden_tour.models.data.ImageRepository;
 import org.launchcode.water_garden_tour.models.data.OwnerRepository;
 import org.launchcode.water_garden_tour.models.garden.Feature;
 import org.launchcode.water_garden_tour.models.garden.Garden;
-import org.launchcode.water_garden_tour.models.garden.Image;
 import org.launchcode.water_garden_tour.models.garden.Owner;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.util.StringUtils;
 import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
-import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
+import javax.servlet.http.HttpServletResponse;
 import javax.validation.Valid;
-import java.io.*;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.nio.file.StandardCopyOption;
-import java.util.Base64;
-import java.util.Objects;
+import java.io.ByteArrayInputStream;
+import java.io.IOException;
+import java.io.InputStream;
 import java.util.Optional;
 
 @Controller
 @RequestMapping("")
 public class GardenController {
-
-    private final String UPLOAD_DIR = "/Users/JamieAndersen/KC-Water-Garden-App/src/main/java/org/launchcode/water_garden_tour/uploads/";
 
     @Autowired
     private GardenRepository gardenRepository;
@@ -41,9 +35,6 @@ public class GardenController {
 
     @Autowired
     private OwnerRepository ownerRepository;
-
-    @Autowired
-    private ImageRepository imageRepository;
 
     //****Garden controllers
 
@@ -57,67 +48,42 @@ public class GardenController {
     }
 
     @PostMapping("/gardens/add")
-    public String addGarden(@ModelAttribute @Valid Garden newGarden,
-                            Errors errors, Model model) {
+    public String addGarden(@ModelAttribute @Valid Garden newGarden, @RequestParam("file") MultipartFile file,
+                            Errors errors, Model model) throws IOException {
 
         if (errors.hasErrors()) {
             return "gardens/add";
         }
 
+        newGarden.setImage(file.getBytes());
+
         gardenRepository.save(newGarden);
         model.addAttribute("gardens", gardenRepository.findAll());
         model.addAttribute("title", "Garden List");
-
-        return "redirect:/gardens/uploadImage";
-    }
-
-    @GetMapping("gardens/uploadImage")
-    public String renderUploadForm(Model model) {
-        model.addAttribute("gardens", gardenRepository.findAll());
-        return "gardens/uploadImage";
-    }
-
-    @PostMapping("gardens/uploadImage")
-    public String uploadFile(@RequestParam("file") MultipartFile file, RedirectAttributes attributes, Image newImage) throws IOException {
-
-        /*// check if file is empty
-        if (file.isEmpty()) {
-            attributes.addFlashAttribute("message", "Please select a file to upload.");
-            return "gardens/uploadImage";
-        }*/
-
-        newImage.setImage(file.getBytes());
-        InputStream inputStream = new ByteArrayInputStream(newImage.getImage());
-
-        imageRepository.save(newImage);
-
-        /*// normalize the file path
-        String fileName = StringUtils.cleanPath(Objects.requireNonNull(file.getOriginalFilename()));
-
-        // save the file on the local file system
-        try {
-            Path path = Paths.get(UPLOAD_DIR + fileName);
-            Files.copy(file.getInputStream(), path, StandardCopyOption.REPLACE_EXISTING);
-
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-
-        // return success response
-        attributes.addFlashAttribute("message", "You successfully uploaded " + fileName + '!');*/
-
 
         return "redirect:/gardens/list";
     }
 
     @GetMapping("/gardens/list")
     public String listGarden(Model model) {
+
         model.addAttribute("gardens", gardenRepository.findAll());
         model.addAttribute("title", "Garden List");
 
         return "gardens/list";
     }
 
+    @GetMapping("/garden/list/{gardenId}")
+    public void showImage(@PathVariable ("gardenId") Integer gardenId,
+                               HttpServletResponse response) throws IOException {
+
+        Optional<Garden> garden = gardenRepository.findById(gardenId);
+
+        response.setContentType("image/jpeg");
+
+        InputStream is = new ByteArrayInputStream(garden.get().getImage());
+        IOUtils.copy(is, response.getOutputStream());
+    }
 
     //****Feature Controllers
