--- src/main/java/org/launchcode/water_garden_tour/images/ImagesService.java
@@ -1,7 +1,0 @@
-package org.launchcode.water_garden_tour.images;
-
-public class ImagesService {
-
-
-
-}
