+++ src/main/java/org/launchcode/water_garden_tour/models/data/GardenRepository.java
@@ -10,6 +10,4 @@ import javax.transaction.Transactional;
 @Transactional
 public interface GardenRepository extends JpaRepository<Garden, Integer> {
 
-    Garden findByName(String name);
-
 }
