--- src/main/resources/templates/gardens/uploadImage.html
@@ -1,30 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org/">
-<head th:replace="fragments :: head">
-    <title></title>
-</head>
-
-<body>
-
-<div th:replace="fragments :: page-header"></div>
-
-<form method="post" enctype="multipart/form-data">
-
-    <div class="form-group">
-        <label th:for="garden">Garden</label>
-        <select name="garden">
-            <option>--Please select one--</option>
-            <option th:each="garden : ${gardens}"
-                    th:text="${garden.name}"
-                    th:value="${garden.id}"></option>
-        </select>
-    </div>
-    <div class="form-group">
-        <label>Upload Image</label>
-        <input class="form-control-file" type="file" name="file"/>
-    </div>
-    <input type="submit" value="Upload Image"/>
-    <input type="submit" value="No image to upload"/>
-</form>
-</body>
-</html>
